fix __init__.py's in tools/package, rename to PackageNamingScheme (was PackagingNamingScheme)
